[MI-3022]:Fixed review comments of PR #256 issue #198
