+++ server/remote/msgraph/get_me.go
@@ -4,15 +4,13 @@
 package msgraph
 
 import (
-	"strings"
-
 	"github.com/mattermost/mattermost-plugin-mscalendar/server/serializer"
 
 	"github.com/pkg/errors"
 )
 
 const (
-	ErrorUserInactive        = "You have been marked inactive. Please disconnect and reconnect your account again."
+	ErrorUserInactive        = "You have been marked inactive because your refresh token is expired. Please disconnect and reconnect your account again."
 	LogUserInactive          = "User %s is inactive. Please disconnect and reconnect your account."
 	ErrorRefreshTokenExpired = "The refresh token has expired due to inactivity"
 )
@@ -44,51 +42,4 @@ func (c *client) GetMe() (*serializer.User, error) {
 	}
 
 	return user, nil
-}
-
-func (c *client) CheckUserStatus() bool {
-	// Not checking for API calls made using super client
-	if c.store == nil {
-		return true
-	}
-
-	user, err := c.store.LoadUser(c.mattermostUserID)
-	if err != nil {
-		c.Logger.Errorf("Not able to load the user %s. %s", c.mattermostUserID, err.Error())
-		return false
-	}
-
-	// Checking if the user is marked as inactive
-	if user.OAuth2Token.AccessToken == "" {
-		if _, err := c.poster.DM(c.mattermostUserID, ErrorUserInactive); err != nil {
-			c.Logger.Errorf("Not able to DM the user %s. %s", c.mattermostUserID, err.Error())
-		}
-		return false
-	}
-
-	return true
-}
-
-func (c *client) ChangeUserStatus(err error) {
-	if !strings.Contains(err.Error(), ErrorRefreshTokenExpired) {
-		return
-	}
-
-	// Not checking for API calls made using super client
-	if c.store == nil {
-		return
-	}
-
-	user, err := c.store.LoadUser(c.mattermostUserID)
-	if err != nil {
-		c.Logger.Errorf("Not able to load the user %s. %s", c.mattermostUserID, err.Error())
-		return
-	}
-
-	// Marking the user as inactive
-	user.OAuth2Token.AccessToken = ""
-	if err = c.store.StoreUser(user); err != nil {
-		c.Logger.Errorf("Not able to store the user %s. %s", c.mattermostUserID, err.Error())
-		return
-	}
 }
